@@ -10,10 +10,12 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0-alpha-81346764b-20210714",
+    "react-dom": "^18.0.0-alpha-81346764b-20210714",
     "react-scripts": "4.0.1",
-    "typescript": "^4.0.3",
+    "styled-components": "^5.2.1",
+    "styled-reset": "^4.3.1",
+    "typescript": "^4.1.2",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
@@ -39,5 +41,13 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@craco/craco": "^5.9.0",
+    "@types/styled-components": "^5.1.4",
+    "autoprefixer": "^9",
+    "postcss": "^7",
+    "prettier": "^2.2.1",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat"
   }
 }
